isola filtro mixin para adição de lookups
